data-down, actions-up only
I'm upgrading an app that used an earlier version of x-toggle. Upgrading Ember to 1.13 exposed timing sensitivities in x-toggle's observers that were causing incorrect upward data flow.

This change makes x-toggle purely DDAU. The `toggled` value comes in, and the `on-toggle` actions goes out.

All of this would be easier to do if I could limit myself to Ember 2.x, but I need a version that will also work back to 1.12.